requirements.txt
playwright
google-generativeai
pytest
pytest-asyncio
